All class now have their own fully functioning views.  These are auto loaded with no extra work as long as the designer of the classes follows the requirements for their behavior, config, and view file.  The plugin will do the rest.
