+++ RPGPlugin.csproj
@@ -127,21 +127,26 @@
       <DependentUpon>Settings.settings</DependentUpon>
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
     </Compile>
-    <Compile Include="RPGPlugin\Roles\Behavior\ClassesBase.cs" />
+    <Compile Include="RPGPlugin\Roles\baseClasses\ClassesBase.cs" />
+    <Compile Include="RPGPlugin\Roles\baseClasses\configBase.cs" />
     <Compile Include="RPGPlugin\Roles\Behavior\HunterClass.cs" />
     <Compile Include="RPGPlugin\Roles\Behavior\MinerClass.cs" />
     <Compile Include="RPGPlugin\Roles\Behavior\SampleClass.cs" />
     <Compile Include="RPGPlugin\Roles\Behavior\WarriorClass.cs" />
-    <Compile Include="RPGPlugin\Roles\Config\configBase.cs" />
     <Compile Include="RPGPlugin\Roles\Config\HunterConfig.cs" />
     <Compile Include="RPGPlugin\Roles\Config\MinerConfig.cs" />
     <Compile Include="RPGPlugin\Roles\Config\PlayerData.cs" />
     <Compile Include="RPGPlugin\Roles\Config\SampleConfig.cs" />
     <Compile Include="RPGPlugin\Roles\Config\WarriorConfig.cs" />
-    <Compile Include="RPGPlugin\Roles\View\HunterView.xaml.cs">
-      <DependentUpon>HunterView.xaml</DependentUpon>
+    <Compile Include="RPGPlugin\Roles\View\Hunter.xaml.cs">
+      <DependentUpon>Hunter.xaml</DependentUpon>
     </Compile>
-    <Compile Include="RPGPlugin\Roles\View\ViewBase.cs" />
+    <Compile Include="RPGPlugin\Roles\View\Miner.xaml.cs">
+      <DependentUpon>Miner.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="RPGPlugin\Roles\View\Warrior.xaml.cs">
+      <DependentUpon>Warrior.xaml</DependentUpon>
+    </Compile>
     <Compile Include="RPGPlugin\RPGPluginManager.cs" />
     <Compile Include="RPGPlugin\RPGPlugin.cs" />
     <Compile Include="RPGPlugin\RPGPluginCommands.cs" />
@@ -184,7 +189,9 @@
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="RPGPlugin\Roles\View\HunterView.xaml" />
+    <Page Include="RPGPlugin\Roles\View\Hunter.xaml" />
+    <Page Include="RPGPlugin\Roles\View\Miner.xaml" />
+    <Page Include="RPGPlugin\Roles\View\Warrior.xaml" />
     <Page Include="RPGPlugin\Torch GUI\EditItemWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
