--- RPGPlugin/Roles/View/HunterView.xaml
@@ -1,12 +1,0 @@
-<UserControl x:Class="RPGPlugin.View.Hunter"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:local="clr-namespace:RPGPlugin.View"
-             mc:Ignorable="d"
-             d:DesignHeight="300" d:DesignWidth="300">
-    <Grid>
-        <TextBlock FontSize="30">Test</TextBlock>
-    </Grid>
-</UserControl>
