+++ RPGPlugin/Roles/View/Miner.xaml
@@ -1,13 +1,13 @@
-﻿<UserControl x:Class="RPGPlugin.RolesControl"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:RPGPlugin"
-      mc:Ignorable="d" 
-      d:DesignHeight="450" d:DesignWidth="800">
-
-    <UserControl.Resources>
+<UserControl x:Class="RPGPlugin.View.Miner"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:local="clr-namespace:RPGPlugin.View"
+             mc:Ignorable="d"
+             d:DesignHeight="300" d:DesignWidth="300">
+    
+     <UserControl.Resources>
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
                 <ResourceDictionary Source="/RPGPlugin;component/CustomStyles.xaml" />
@@ -75,45 +75,22 @@
             </ControlTemplate>
         </ResourceDictionary>
     </UserControl.Resources>
-
+    
     <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="Auto"/>
-        </Grid.ColumnDefinitions>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="*"/>
-        </Grid.RowDefinitions>
-        <TabControl Name="SettingsTab">
-            <TabItem Header="Main" Style="{StaticResource TabItemStyle}" IsSelected="True">
-                <Grid>
-                    <Grid.ColumnDefinitions>
-                        <ColumnDefinition Width="20"/> <!-- Spacer to make easy aligning without margins -->
-                        <ColumnDefinition Width="200"/>
-                        <ColumnDefinition Width="400"/>
-                        <ColumnDefinition Width="*"/>
-                        <ColumnDefinition Width="20"/>
-                    </Grid.ColumnDefinitions>
-                    <Grid.RowDefinitions>
-                        <RowDefinition Height="Auto"/>
-                        <RowDefinition Height="Auto"/>
-                        <RowDefinition Height="Auto"/>
-                        <RowDefinition Height="Auto"/>
-                    </Grid.RowDefinitions>
+        <Button x:Name="EditOreButton" Content="EDIT" Style="{StaticResource ButtonStyle}" Margin="115,10,0,0" HorizontalAlignment="Left" VerticalAlignment="Top" Grid.Column="0" Grid.Row="0" Click="EditOreButton_Click" Width="75" Height="25" FontSize="10" FontWeight="Bold"/>
+        <Button x:Name="DeleteOreButton" Content="DELETE" Style="{StaticResource ButtonStyle}" Margin="195,10,0,0" HorizontalAlignment="Left" VerticalAlignment="Top" Grid.Column="0" Grid.Row="0" Click="DeleteOreButton_Click" Width="75" Height="25" FontSize="10" FontWeight="Bold"/>
 
-                    <StackPanel Grid.Column="0" Grid.Row="0" Grid.ColumnSpan="6" Orientation="Horizontal" HorizontalAlignment="Center">
-                        <Button Style="{StaticResource DonateButtonStyle}" Click="Donate_Click" Margin="-130,0,0,10" />
-                        <Button Height="64" Width="64" Style="{StaticResource supportbutton}" Click="SupportButton_Click" Margin="10"/>
-                    </StackPanel>
-                    
-                    <TextBlock Grid.Column="1" Grid.Row="1" Text="NOTIFICATIONS LVLUP" VerticalAlignment="Center" Margin="5"/>
-                    <ToggleButton Grid.Column="2" Grid.Row="1" HorizontalAlignment="Left" x:Name="BroadcastLevelUpToggleButton" Style="{StaticResource ToggleButtonStyle1}" Checked="BroadcastLevelUpToggleButton_Checked" Unchecked="BroadcastLevelUpToggleButton_Unchecked" Margin="5"/>
-                    
-                    <TextBlock Grid.Column="1" Grid.Row="2" Text="Plugin Base Save Location" Margin="5"/>
-                    <TextBox Name="BaseSaveLocation" Grid.Column="2" Grid.Row="2" Margin="5" Text="{Binding SaveLocation}"/>
-                    <TextBlock Grid.Column="1" Grid.Row="3" Grid.ColumnSpan="2" Text="This is where all Class configurations and player data will be saved.  Using a network location allows this to be used across multiple instances (Nexus)." TextWrapping="WrapWithOverflow"/>
+        <Button x:Name="AddNewOreButton" Content="ADD NEW ORE" Style="{StaticResource ButtonStyle}" Margin="10,10,0,0" HorizontalAlignment="Left" VerticalAlignment="Top" Grid.Column="0" Grid.Row="0" Click="AddNewOreButton_Click" Background="DarkSlateGray" Width="100" Height="25" FontWeight="Bold" FontSize="10"/>
 
-                </Grid>
-            </TabItem>
-        </TabControl>
+        <DataGrid Name="ExpRatioDataGrid" CanUserAddRows="False" CanUserDeleteRows="False" SelectionMode="Single"
+                  Height="372" VerticalAlignment="Top" AutoGenerateColumns="False"
+                  Margin="10,39,0,-1" AlternatingRowBackground="Snow" HorizontalAlignment="Left" Width="260" Style="{StaticResource DataGridStyle}" Background="#FFF0F0F0" FontWeight="Bold" >
+            <DataGrid.Columns>
+                <DataGridTextColumn Header="MINERAL" Binding="{Binding Path=Key}" IsReadOnly="True" Width="*" HeaderStyle="{StaticResource DataGridColumnHeaderStyle}" />
+                <DataGridTextColumn Header="EXPERIENCE" Binding="{Binding Path=Value}" IsReadOnly="True" Width="*" HeaderStyle="{StaticResource DataGridColumnHeaderStyle2}" />
+            </DataGrid.Columns>
+        </DataGrid>
+
+
     </Grid>
 </UserControl>
