--- RPGPlugin/Roles/View/ViewBase.cs
@@ -1,10 +1,0 @@
-using System.Windows.Controls;
-
-namespace RPGPlugin.View
-{
-    public interface ViewBase
-    {
-        void LogThread();
-
-    }
-}
