+++ RPGPlugin/Roles/baseClasses/ClassesBase.cs
